tests for planned workouts
