--- boot/src/test/groovy/org/freekode/tp2intervals/infrastructure/intervalsicu/workout/IntervalsWorkoutRepositoryTest.groovy
@@ -1,35 +1,0 @@
-package org.freekode.tp2intervals.infrastructure.intervalsicu.workout
-
-import org.freekode.tp2intervals.config.ISpringConfiguration
-import org.freekode.tp2intervals.domain.config.AppConfigurationRepository
-import org.freekode.tp2intervals.domain.config.UpdateConfigurationRequest
-import org.springframework.beans.factory.annotation.Autowired
-import org.springframework.boot.test.context.SpringBootTest
-import org.springframework.test.context.ActiveProfiles
-
-import java.time.LocalDate
-
-class IntervalsWorkoutRepositoryTest extends ISpringConfiguration {
-    @Autowired
-    IntervalsWorkoutRepository intervalsWorkoutRepository
-
-    @Autowired
-    AppConfigurationRepository appConfigurationRepository
-
-    def setup() {
-        appConfigurationRepository.updateConfig(new UpdateConfigurationRequest(
-                Map.of(
-                        "intervals.api-key", "my-key",
-                        "intervals.athlete-id", "my-athlete"
-                )
-        ))
-    }
-
-    def "should parse and map planned workouts"() {
-        when:
-        def workouts = intervalsWorkoutRepository.getPlannedWorkouts(LocalDate.now(), LocalDate.now())
-
-        then:
-        workouts != null
-    }
-}
