+++ boot/src/test/groovy/org/freekode/tp2intervals/infrastructure/trainerroad/activity/TrainerRoadActivityRepositoryIT.groovy
@@ -7,7 +7,7 @@ import org.springframework.beans.factory.annotation.Autowired
 
 import java.time.LocalDate
 
-class TrainerRoadActivityRepositoryTest extends ISpringConfiguration {
+class TrainerRoadActivityRepositoryIT extends ISpringConfiguration {
     @Autowired
     TrainerRoadActivityRepository trainerRoadActivityRepository
