+++ boot/src/test/resources/intervals-events-response-note.json
@@ -1,0 +1,52 @@
+[
+  {
+    "id": 31115158,
+    "start_date_local": "2024-02-05T00:00:00",
+    "icu_training_load": null,
+    "icu_atl": null,
+    "icu_ctl": null,
+    "type": null,
+    "calendar_id": 1,
+    "uid": "d85038d2-b5d1-4d94-afbf-2419f850df8e",
+    "athlete_id": "i11111",
+    "category": "NOTE",
+    "end_date_local": "2024-02-06T00:00:00",
+    "name": "note test",
+    "description": "",
+    "indoor": false,
+    "color": null,
+    "moving_time": null,
+    "icu_ftp": null,
+    "atl_days": null,
+    "ctl_days": null,
+    "updated": "2024-02-05T11:08:56.819+00:00",
+    "not_on_fitness_chart": false,
+    "show_as_note": false,
+    "show_on_ctl_line": false,
+    "target": null,
+    "joules": null,
+    "joules_above_ftp": null,
+    "workout_doc": null,
+    "push_errors": null,
+    "athlete_cannot_edit": false,
+    "hide_from_athlete": false,
+    "structure_read_only": false,
+    "created_by_id": "i11111",
+    "shared_event_id": null,
+    "entered": false,
+    "tags": null,
+    "attachments": null,
+    "oauth_client_id": null,
+    "external_id": null,
+    "load_target": null,
+    "time_target": null,
+    "distance_target": null,
+    "plan_athlete_id": null,
+    "plan_folder_id": null,
+    "plan_workout_id": null,
+    "plan_applied": null,
+    "distance": null,
+    "icu_intensity": null,
+    "plan_name": null
+  }
+]
